requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
